+++ src/plugins/index.ts
@@ -93,6 +93,3 @@ export {GLTFKHRMaterialVariantsPlugin} from './extras/GLTFKHRMaterialVariantsPlu
 export {DynamicImportPlugin, sampleThreepipeViteHmrPlugin} from './extras/DynamicImportPlugin'
 export {ObjectConstraintsPlugin, ObjectConstraint, type ObjectConstraintsPluginEventMap} from './extras/ObjectConstraintsPlugin'
 export {basicObjectConstraints, type ConstraintPropsTypes, type TConstraintPropsType, type ConstraintPropsType} from './extras/helpers/BasicObjectConstraints'
-export * /* {EntityComponentPlugin} */from './extras/EntityComponentPlugin'
-export {CannonPhysicsPlugin} from './extras/CannonPhysicsPlugin'
-export {SkyPlugin} from './extras/atmosphere/SkyPlugin'