+++ src/utils/serialization.ts
@@ -470,6 +470,12 @@ export class ThreeSerialization {
         // Serialization.SerializableClasses.set('Skeleton', Skeleton) // doesnt have .type. todo add to three.js
     }
 
+    static MakeSerializable(constructor: ObjectConstructor, type: string, props: (string|[string, string])[]) {
+        (constructor.prototype as any).serializableClassId = type
+        Serialization.SerializableClasses.set(type, constructor)
+        Serialization.TypeMap.set(constructor, props.map(p=>typeof p === 'string' ? [p, p] : p))
+    }
+
     /**
      * Serialize an object
      * {@link Serialization.Serialize}